Abort translation on parse errors
For requests, we should abort translating key format if a parsing error
occurs due to invalid json. Let the request continue untouched, similar
to the scenario when content-type header is missing

It shouldn't be this middleware's responcibility to deal with invalid
json. The app server can handle these scenarios on a case by case basis

https://www.pivotaltracker.com/story/show/176519829
[#176519829]